infra/requirements.txt
python-dotenv==1.0.1
pulumi-docker-build==0.0.7
PyGithub==2.5.0
pulumi==3.136.1
pulumi-azure-native==2.66.0
pulumi-azure==6.5.0
pulumi_random==4.16.7
pulumi_command==1.0.1
pulumi_azuread==6.0.1
PyNaCl==1.5.0
